Change project structure
Add ruleset, stylecop, editorconfig
Refactor code
Move enum to separate file
