+++ .editorconfig
@@ -1,0 +1,120 @@
+# To learn more about .editorconfig see https://docs.microsoft.com/en-us/visualstudio/ide/editorconfig-code-style-settings-reference
+
+root = true
+
+# All files
+[*]
+indent_style = space
+indent_size = 2
+insert_final_newline = true:error
+trim_trailing_whitespace = true:error
+guidelines = 120
+
+# Xml files
+[*.{xml,xaml,config,aspx,cshtml,html}]
+indent_size = 2
+
+# C# files
+[*.cs]
+
+#'this.' and 'Me.' qualifiers
+dotnet_style_qualification_for_field = false:error
+dotnet_style_qualification_for_method = false:error
+dotnet_style_qualification_for_property = false:error
+dotnet_style_qualification_for_event = false:error
+
+#Predefined types (int, string, etc.)
+dotnet_style_predefined_type_for_locals_parameters_members = true:error
+dotnet_style_predefined_type_for_member_access = false:error
+
+#Modifiers
+dotnet_style_require_accessibility_modifiers = never:error
+dotnet_style_readonly_field = true:error
+
+#Parentheses 
+dotnet_style_parentheses_in_arithmetic_binary_operators = always_for_clarity:refactoring
+dotnet_style_parentheses_in_other_binary_operators = always_for_clarity:refactoring
+dotnet_style_parentheses_in_relational_binary_operators = always_for_clarity:refactoring
+dotnet_style_parentheses_in_other_operators = nenever_if_unnecessary:error
+
+#Expressions
+dotnet_style_object_initializer = true:error
+dotnet_style_collection_initializer = true:error
+dotnet_style_explicit_tuple_names = true:error
+dotnet_style_prefer_inferred_tuple_names = true:error
+dotnet_style_prefer_inferred_anonymous_type_member_names = true:error
+dotnet_style_prefer_auto_properties = true:error
+dotnet_style_prefer_is_null_check_over_reference_equality_method = true:error
+dotnet_style_prefer_conditional_expression_over_assignment = true:error
+dotnet_style_prefer_conditional_expression_over_return = true:suggestion
+
+#Null-checking
+dotnet_style_coalesce_expression = true:error
+dotnet_style_null_propagation = true:error
+csharp_style_throw_expression = false:error
+csharp_style_conditional_delegate_call = true:error
+
+#Implicit and explicit types
+csharp_style_var_for_built_in_types = false:error
+csharp_style_var_when_type_is_apparent = true:error
+csharp_style_var_elsewhere = true:error
+
+#Expression-bodied members
+csharp_style_expression_bodied_methods = false:error
+csharp_style_expression_bodied_constructors = false:error
+csharp_style_expression_bodied_operators = true:error
+csharp_style_expression_bodied_properties = true:error
+csharp_style_expression_bodied_indexers = true:error
+csharp_style_expression_bodied_accessors = true:error
+
+#Pattern matching for 'is' and 'as'
+csharp_style_pattern_matching_over_is_with_cast_check = true:error
+csharp_style_pattern_matching_over_as_with_null_check = true:error
+
+#Inlined variable declarations for 'out'
+csharp_style_inlined_variable_declaration = true:error
+
+#Expression-level
+csharp_prefer_simple_default_expression = true:suggestion
+csharp_style_deconstructed_variable_declaration = true:suggestion
+
+#Code block braces
+csharp_prefer_braces = true:error
+
+#Usings
+dotnet_sort_system_directives_first = true
+dotnet_separate_import_directive_groups = false
+
+
+#New line
+csharp_new_line_before_open_brace = all #control_blocks, lambdas, properties, object_collection_array_initializers, types, accessors, methods
+csharp_new_line_before_else = true
+csharp_new_line_before_catch = true
+csharp_new_line_before_finally = true
+csharp_new_line_before_members_in_object_initializers = true
+csharp_new_line_before_members_in_anonymous_types = true
+csharp_new_line_between_query_expression_clauses = true
+
+#Indentation
+csharp_indent_case_contents = true
+csharp_indent_case_contents_when_block = true
+csharp_indent_switch_labels = true
+csharp_indent_labels = no_change
+
+
+#Spacing
+csharp_space_after_cast = false
+csharp_space_after_keywords_in_control_flow_statements = true
+csharp_space_between_method_declaration_parameter_list_parentheses = false
+csharp_space_between_method_call_parameter_list_parentheses = false
+csharp_space_between_parentheses = false
+csharp_space_before_colon_in_inheritance_clause = false
+csharp_space_after_colon_in_inheritance_clause = true
+csharp_space_around_binary_operators = before_and_after
+csharp_space_between_method_declaration_empty_parameter_list_parentheses = false
+csharp_space_between_method_call_name_and_opening_parenthesis = false
+csharp_space_between_method_call_empty_parameter_list_parentheses = false
+
+#Wrapping options
+csharp_preserve_single_line_statements = false
+csharp_preserve_single_line_blocks = true
