+++ Patients/Enums/Sex.cs
@@ -1,0 +1,8 @@
+﻿namespace Patients.Enums
+{
+  public enum Sex
+  {
+    Male,
+    Female
+  }
+}
